requirements.txt
blinker            == 1.9.0
certifi            == 2025.1.31
charset-normalizer == 3.4.1
click              == 8.1.8
dataclasses        == 0.6
dnslib             == 0.9.26
Flask              == 3.1.0
idna               == 3.10
itsdangerous       == 2.2.0
Jinja2             == 3.1.6
jsonify            == 0.5
MarkupSafe         == 3.0.2
paho-mqtt          == 2.1.0
pyderive3          == 0.0.8
pydhcp3            == 0.0.6
pypool3            == 0.0.2
pyserve            == 0.2.8
pyserve3           == 0.0.7
pystructs3         == 0.0.8
requests           == 2.32.3
typing_extensions  == 4.12.2
urllib3            == 2.3.0
Werkzeug           == 3.1.3
